added the ability to send the friend request notification

+++ server-Chat-App/src/main/java/org/server/chatapp/dao/implement/UsersImpl.java
@@ -178,7 +178,7 @@ public class UsersImpl implements UsersDao {
     }
 
     @Override
-    public int insert(Users users) {
+    public long insert(Users users) {
         int result = 0;
         try (Connection connection = Database.getDataSource().getConnection()) {
             String sql = """
